port dl4j matrix calculations to new nd4j arch
Former-commit-id: dd7f875a30df5f7b9606f10f00bb58250f0be3d7

+++ deeplearning4j-core/src/main/java/org/deeplearning4j/models/featuredetectors/autoencoder/AutoEncoder.java
@@ -61,7 +61,7 @@ public class AutoEncoder extends BaseNeuralNetwork {
 
     @Override
     public NeuralNetworkGradient getGradient(Object[] params) {
-        float lr = (float) params[0];
+        double lr = (double) params[0];
         int iterations = (int) params[1];
