+++ deeplearning4j-core/src/main/java/org/deeplearning4j/optimize/api/LineOptimizerMatrix.java
@@ -10,10 +10,10 @@ import org.nd4j.linalg.api.ndarray.INDArray;
  */
 public interface LineOptimizerMatrix {
 	/** Returns the last step size used. */
-	public float optimize (INDArray line, int iteration,float initialStep) throws InvalidStepException;
+	public double optimize (INDArray line, int iteration,double initialStep) throws InvalidStepException;
 
 	public interface ByGradient	{
 		/** Returns the last step size used. */
-		public float optimize (INDArray line, int iteration,float initialStep) throws InvalidStepException;
+		public double optimize (INDArray line, int iteration,double initialStep) throws InvalidStepException;
 	}
 }
